# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AnimalHealthConditionRequestDTO" : {
        "properties" : {
          "description" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "ALERGIA", "CONDICAO_CRONICA" ],
            "type" : "string"
          },
          "notes" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "resolvedAt" : {
            "format" : "date",
            "type" : "string"
          },
          "severity" : {
            "enum" : [ "LEVE", "MODERADA", "GRAVE" ],
            "type" : "string"
          },
          "since" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "description", "kind" ],
        "type" : "object"
      },
      "AnimalHealthConditionResponseDTO" : {
        "properties" : {
          "animalHealthConditionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ativa" : {
            "type" : "boolean"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "ALERGIA", "CONDICAO_CRONICA" ],
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "resolvedAt" : {
            "format" : "date",
            "type" : "string"
          },
          "severity" : {
            "enum" : [ "LEVE", "MODERADA", "GRAVE" ],
            "type" : "string"
          },
          "since" : {
            "format" : "date",
            "type" : "string"
          },
          "updateDate" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "AntiparasiticRequestDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "antiparasiticCatalogId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "applicationDate" : {
            "format" : "date",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "DEWORMER", "FLEA_TICK" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "nextDoseDate" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "animalId" ],
        "type" : "object"
      },
      "AntiparasiticResponseDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "antiparasiticCatalogId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "antiparasiticId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "applicationDate" : {
            "format" : "date",
            "type" : "string"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "DEWORMER", "FLEA_TICK" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "nextDoseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "updateDate" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "AttachmentResponseDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "attachmentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "checksumSha256" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "healthRecordId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "originalFilename" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "format" : "int64",
            "type" : "integer"
          },
          "uploadedByPersonName" : {
            "type" : "string"
          },
          "vaccineId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CareInstructionRequestDTO" : {
        "properties" : {
          "description" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "endsOn" : {
            "format" : "date",
            "type" : "string"
          },
          "intervalDays" : {
            "format" : "int32",
            "type" : "integer"
          },
          "startsOn" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "description", "intervalDays", "startsOn" ],
        "type" : "object"
      },
      "CareInstructionResponseDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "careInstructionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "creationDate" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "endsOn" : {
            "format" : "date",
            "type" : "string"
          },
          "intervalDays" : {
            "format" : "int32",
            "type" : "integer"
          },
          "lastFulfilledAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lastFulfilledByName" : {
            "type" : "string"
          },
          "organizationName" : {
            "type" : "string"
          },
          "recordedByName" : {
            "type" : "string"
          },
          "revokedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "startsOn" : {
            "format" : "date",
            "type" : "string"
          },
          "vigente" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ObservationRequestDTO" : {
        "properties" : {
          "description" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "observedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "urgent" : {
            "type" : "boolean"
          }
        },
        "required" : [ "description" ],
        "type" : "object"
      },
      "ObservationResponseDTO" : {
        "properties" : {
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "observationId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "observedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "organizationName" : {
            "type" : "string"
          },
          "recordedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "recordedByName" : {
            "type" : "string"
          },
          "urgent" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "PageSensitiveAccessLogResponseDTO" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/SensitiveAccessLogResponseDTO"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "SensitiveAccessLogResponseDTO" : {
        "properties" : {
          "accessedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "actorName" : {
            "type" : "string"
          },
          "actorType" : {
            "enum" : [ "VET", "SHARE_LINK" ],
            "type" : "string"
          },
          "animalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ipAddress" : {
            "type" : "string"
          },
          "organizationName" : {
            "type" : "string"
          },
          "resource" : {
            "enum" : [ "VACCINES", "HEALTH_RECORDS", "VACCINE_CORRECTIONS", "HEALTH_RECORD_CORRECTIONS", "ATTACHMENTS", "SHARED_CARD" ],
            "type" : "string"
          },
          "sensitiveAccessLogId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "ascending" : {
            "type" : "boolean"
          },
          "direction" : {
            "type" : "string"
          },
          "ignoreCase" : {
            "type" : "boolean"
          },
          "nullHandling" : {
            "type" : "string"
          },
          "property" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "O token devolvido por POST /auth/login.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/animals/{animalId}/access-log" : {
      "get" : {
        "description" : "Do mais recente para o mais antigo. Inclui leitura por veterinario de clinica autorizada e abertura do link publico de carteira. Leitura do proprio tutor e dos co-tutores nao entra - o log responde 'quem MAIS viu isto', e a resposta nao inclui quem pergunta.",
        "operationId" : "listByAnimal_6",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageSensitiveAccessLogResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Acessos de terceiros ao dado de saude do animal",
        "tags" : [ "sensitive-access-log-controller" ]
      }
    },
    "/animals/{animalId}/attachments" : {
      "get" : {
        "description" : "Sem filtro, devolve todos os anexos do animal. Com vaccineId ou healthRecordId, devolve os daquele registro.",
        "operationId" : "listByAnimal_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "vaccineId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "healthRecordId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AttachmentResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Anexos do animal",
        "tags" : [ "attachment-controller" ]
      },
      "post" : {
        "description" : "Multipart. Aceita JPEG, PNG, WEBP e PDF, reconhecidos pelo CONTEUDO - o Content-Type declarado nao e consultado. vaccineId e healthRecordId sao opcionais e mutuamente exclusivos: dizem o que o arquivo documenta. Sem nenhum dos dois, o anexo e do animal em si.",
        "operationId" : "upload",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "vaccineId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "healthRecordId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "description",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttachmentResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Anexa um arquivo ao animal",
        "tags" : [ "attachment-controller" ]
      }
    },
    "/animals/{animalId}/care-instructions" : {
      "get" : {
        "description" : "Inclui as encerradas: tratamento suspenso e informacao clinica, e quem ler o historico depois precisa saber que houve. O campo vigente diz o que vale hoje.",
        "operationId" : "listByAnimal_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CareInstructionResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Orientacoes do animal",
        "tags" : [ "care-instruction-controller" ]
      },
      "post" : {
        "description" : "Serve a prescricao do veterinario, a medicacao continua e o tema de casa da creche - os tres tem a mesma forma. Informe X-Petfy-Organization para emitir em nome de uma organizacao; sem o header a orientacao sai assinada apenas pela pessoa. endsOn ausente significa tratamento indefinido.",
        "operationId" : "create_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CareInstructionRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CareInstructionResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Emite uma orientacao de cuidado",
        "tags" : [ "care-instruction-controller" ]
      }
    },
...
    "/animals/{animalId}/conditions" : {
      "get" : {
        "description" : "As ativas vem primeiro: a pergunta que a lista responde e o que vale para este animal hoje. Condicao encerrada aparece com resolvedAt preenchido, porque faz parte do historico.",
        "operationId" : "listByAnimal_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AnimalHealthConditionResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Alergias e condicoes do animal",
        "tags" : [ "animal-health-condition-controller" ]
      },
      "post" : {
        "description" : "severity vale apenas para ALERGIA. Em CONDICAO_CRONICA responde 400 - gravidade de diabetes nao se mede em leve, moderada e grave.",
        "operationId" : "create_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnimalHealthConditionRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnimalHealthConditionResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra alergia ou condicao cronica",
        "tags" : [ "animal-health-condition-controller" ]
      }
    },
...
    "/animals/{animalId}/observations" : {
      "get" : {
        "description" : "Ordenadas por quando foram vistas, mais recentes primeiro - e nao por quando foram digitadas.",
        "operationId" : "listByAnimal_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ObservationResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "As observacoes do animal",
        "tags" : [ "observation-controller" ]
      },
      "post" : {
        "description" : "Qualquer um com acesso de escrita registra: tutor, monitor, lar transitorio, voluntario. observedAt ausente assume agora - preencha quando o que se viu aconteceu antes da digitacao. urgent sinaliza o alerta, que pede atencao sem ser emergencia medica nem ato clinico.",
        "operationId" : "create_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ObservationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ObservationResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra o que se viu com o animal",
        "tags" : [ "observation-controller" ]
      }
    },
...
    "/antiparasitics" : {
      "get" : {
        "description" : "animalId e obrigatorio: a listagem e sempre escopada ao animal, para nao misturar vermifugo de animais diferentes numa lista plana.",
        "operationId" : "listByAnimal_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "animalId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AntiparasiticResponseDTO"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "As aplicacoes de antiparasitario de um animal",
        "tags" : [ "antiparasitic-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AntiparasiticRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AntiparasiticResponseDTO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra uma aplicacao de antiparasitario",
        "tags" : [ "antiparasitic-controller" ]
      }
    },
...
    "/attachments/{attachmentId}" : {
      "delete" : {
        "description" : "Apaga a linha e os BYTES no disco. Arquivo orfao com laudo dentro e dado pessoal nao apagado, que e o oposto do que um pedido de exclusao pede.",
        "operationId" : "delete",
        "parameters" : [ {
          "in" : "path",
          "name" : "attachmentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Apaga o anexo",
        "tags" : [ "attachment-controller" ]
      }
    },
    "/attachments/{attachmentId}/content" : {
      "get" : {
        "description" : "Autorizado a cada chamada pelo animal do anexo. Nao ha URL assinada: link encaminhado por engano daria acesso a dado de saude ate expirar.",
        "operationId" : "download",
        "parameters" : [ {
          "in" : "path",
          "name" : "attachmentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Baixa o conteudo do anexo",
        "tags" : [ "attachment-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ]
}